feat: add ci/cd github actions workflow
