+++ backend/src/server.ts
@@ -1,38 +1,6 @@
-import express, { Request, Response } from 'express';
-import cors from 'cors';
-import dotenv from 'dotenv';
+import app from './app';
 
-// Import route modules
-import authRoutes from './routes/auth';
-import protectedRoutes from './routes/protected';
-import healthRoutes from './routes/health';
-
-dotenv.config();
-
-const app = express();
 const PORT = process.env.PORT || 3001;
-
-// Middleware
-app.use(cors());
-app.use(express.json());
-
-// API Routes
-app.use('/api/auth', authRoutes);
-app.use('/api/protected', protectedRoutes);
-app.use('/api/health', healthRoutes);
-
-// Root endpoint
-app.get('/', (req: Request, res: Response) => {
-    res.json({
-        message: 'Web3 Authentication API',
-        version: '1.0.0',
-        endpoints: {
-            auth: '/api/auth',
-            protected: '/api/protected',
-            health: '/api/health'
-        }
-    });
-});
 
 app.listen(PORT, () => {
     console.log(`🚀 Backend server running on port ${PORT}`);
