+++ backend/vercel.json
@@ -1,0 +1,6 @@
+{
+  "$schema": "https://openapi.vercel.sh/vercel.json",
+  "installCommand": "cd .. && pnpm install --frozen-lockfile",
+  "buildCommand": "pnpm build",
+  "rewrites": [{ "source": "/(.*)", "destination": "/api" }]
+}
